+++ Mini05_Watch Watch App/Model/MetricsTimelineSchedule.swift
@@ -1,0 +1,29 @@
+//
+//  MetricsTimelineSchedule.swift
+//  Mini05_Watch Watch App
+//
+//  Created by Gustavo Horestee Santos Barros on 27/05/24.
+//
+
+import SwiftUI
+
+struct MetricsTimelineSchedule: TimelineSchedule {
+    var startDate: Date
+    var isPaused: Bool
+
+    init(from startDate: Date, isPaused: Bool) {
+        self.startDate = startDate
+        self.isPaused = isPaused
+    }
+
+    func entries(from startDate: Date, mode: TimelineScheduleMode) -> AnyIterator<Date> {
+        var baseSchedule = PeriodicTimelineSchedule(from: self.startDate,
+                                                    by: (mode == .lowFrequency ? 1.0 : 1.0 / 30.0))
+            .entries(from: startDate, mode: mode)
+        
+        return AnyIterator<Date> {
+            guard !isPaused else { return nil }
+            return baseSchedule.next()
+        }
+    }
+}
