+++ Mini05_Watch Watch App/View/Screens/ContentView.swift
@@ -45,5 +45,6 @@ extension ContentView{
 
 #Preview {
     ContentView()
+        .environmentObject(HealthKitManager())
 }
