+++ Mini05_Watch Watch App/View/Screens/SelectEvaluation/SelectEvaluationView.swift
@@ -41,5 +41,6 @@ struct SelectEvaluationView: View {
 
 #Preview {
     SelectEvaluationView()
+        .environmentObject(HealthKitManager())
 }
