+++ Mini05_Watch Watch App/View/Screens/WorkoutInProgress/TabViewWorkout.swift
@@ -30,5 +30,6 @@ struct TabViewWorkout: View {
 
 #Preview {
     TabViewWorkout()
+        .environmentObject(HealthKitManager())
 }
